+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Refresh orders
 - Order orders by timestamp
 - Driver's notification when order is ready
+- Add Minimum Purchase Charge
 
 # v0.2.0
 - Add `ready` state to orders